Move `c8y_http_proxy` under `crates/extensions`
