+++ crates/bin/c8y-device-management/src/config_manager/actor.rs
@@ -10,7 +10,6 @@ use super::upload::ConfigUploadManager;
 use super::upload::UploadConfigFileStatusMessage;
 use super::ConfigManagerConfig;
 use super::DEFAULT_PLUGIN_CONFIG_FILE_NAME;
-use crate::c8y_http_proxy::handle::C8YHttpProxy;
 use anyhow::Result;
 use async_trait::async_trait;
 use c8y_api::smartrest::smartrest_deserializer::SmartRestConfigDownloadRequest;
@@ -18,6 +17,7 @@ use c8y_api::smartrest::smartrest_deserializer::SmartRestConfigUploadRequest;
 use c8y_api::smartrest::smartrest_deserializer::SmartRestRequestGeneric;
 use c8y_api::smartrest::smartrest_serializer::TryIntoOperationStatusMessage;
 use c8y_api::smartrest::topic::C8yTopic;
+use c8y_http_proxy::handle::C8YHttpProxy;
 use mqtt_channel::Message;
 use mqtt_channel::Topic;
 use tedge_actors::fan_in_message_type;
