+++ crates/bin/c8y-device-management/src/config_manager/tests.rs
@@ -1,13 +1,13 @@
-use crate::c8y_http_proxy::messages::C8YRestRequest;
-use crate::c8y_http_proxy::messages::C8YRestResponse;
-use crate::c8y_http_proxy::messages::C8YRestResult;
-use crate::c8y_http_proxy::messages::DownloadFile;
-use crate::c8y_http_proxy::messages::UploadConfigFile;
 use crate::config_manager::actor::OperationTimeout;
 use crate::config_manager::actor::OperationTimer;
 use crate::config_manager::ConfigManagerBuilder;
 use crate::config_manager::ConfigManagerConfig;
 use c8y_api::smartrest::topic::C8yTopic;
+use c8y_http_proxy::messages::C8YRestRequest;
+use c8y_http_proxy::messages::C8YRestResponse;
+use c8y_http_proxy::messages::C8YRestResult;
+use c8y_http_proxy::messages::DownloadFile;
+use c8y_http_proxy::messages::UploadConfigFile;
 use mqtt_channel::Topic;
 use std::time::Duration;
 use tedge_actors::Actor;
