+++ crates/extensions/c8y_http_proxy/src/actor.rs
@@ -1,14 +1,14 @@
-use crate::c8y_http_proxy::credentials::JwtRequest;
-use crate::c8y_http_proxy::credentials::JwtResult;
-use crate::c8y_http_proxy::credentials::JwtRetriever;
-use crate::c8y_http_proxy::messages::C8YRestError;
-use crate::c8y_http_proxy::messages::C8YRestRequest;
-use crate::c8y_http_proxy::messages::C8YRestResult;
-use crate::c8y_http_proxy::messages::DownloadFile;
-use crate::c8y_http_proxy::messages::EventId;
-use crate::c8y_http_proxy::messages::Unit;
-use crate::c8y_http_proxy::messages::UploadConfigFile;
-use crate::c8y_http_proxy::messages::UploadLogBinary;
+use crate::credentials::JwtRequest;
+use crate::credentials::JwtResult;
+use crate::credentials::JwtRetriever;
+use crate::messages::C8YRestError;
+use crate::messages::C8YRestRequest;
+use crate::messages::C8YRestResult;
+use crate::messages::DownloadFile;
+use crate::messages::EventId;
+use crate::messages::Unit;
+use crate::messages::UploadConfigFile;
+use crate::messages::UploadLogBinary;
 use crate::C8YHttpConfig;
 use async_trait::async_trait;
 use c8y_api::http_proxy::C8yEndPoint;
