+++ crates/extensions/c8y_http_proxy/src/credentials.rs
