+++ crates/extensions/c8y_http_proxy/src/lib.rs
@@ -1,8 +1,8 @@
-use crate::c8y_http_proxy::actor::C8YHttpProxyMessageBox;
-use crate::c8y_http_proxy::credentials::JwtResult;
-use crate::c8y_http_proxy::credentials::JwtRetriever;
-use crate::c8y_http_proxy::messages::C8YRestRequest;
-use crate::c8y_http_proxy::messages::C8YRestResult;
+use crate::actor::C8YHttpProxyMessageBox;
+use crate::credentials::JwtResult;
+use crate::credentials::JwtRetriever;
+use crate::messages::C8YRestRequest;
+use crate::messages::C8YRestResult;
 use std::convert::Infallible;
 use std::path::PathBuf;
 use tedge_actors::Builder;
